+++ src/abacus-backoffice/pages/products/add-ons.js
@@ -4,6 +4,6 @@ import * as React from 'react';
 
 import ProductsAddonsLayout from '../../src/products/ProductsAddonsLayout';
 
-export default function ProductsPage(): React.Node {
+export default function ProductAddonsPage(): React.Node {
   return <ProductsAddonsLayout />;
 }