+++ src/abacus-backoffice/src/products/__generated__/ProductsCardsData.graphql.js
@@ -1,0 +1,119 @@
+/**
+ * @flow
+ */
+
+/* eslint-disable */
+
+import type { ReaderFragment } from 'relay-runtime';
+export type SupportedCurrency = "MXN" | "%future added value";
+import type { FragmentReference } from "relay-runtime";
+declare export opaque type ProductsCardsData$ref: FragmentReference;
+declare export opaque type ProductsCardsData$fragmentType: ProductsCardsData$ref;
+export type ProductsCardsData = $ReadOnlyArray<{|
+  +id: string,
+  +key: string,
+  +name: string,
+  +imageCover: ?{|
+    +blurhash: string,
+    +url: string,
+  |},
+  +price: {|
+    +unitAmount: number,
+    +unitAmountCurrency: SupportedCurrency,
+  |},
+  +$refType: ProductsCardsData$ref,
+|}>;
+export type ProductsCardsData$data = ProductsCardsData;
+export type ProductsCardsData$key = $ReadOnlyArray<{
+  +$data?: ProductsCardsData$data,
+  +$fragmentRefs: ProductsCardsData$ref,
+  ...
+}>;
+
+
+const node: ReaderFragment = {
+  "argumentDefinitions": [],
+  "kind": "Fragment",
+  "metadata": {
+    "plural": true
+  },
+  "name": "ProductsCardsData",
+  "selections": [
+    {
+      "alias": null,
+      "args": null,
+      "kind": "ScalarField",
+      "name": "id",
+      "storageKey": null
+    },
+    {
+      "alias": null,
+      "args": null,
+      "kind": "ScalarField",
+      "name": "key",
+      "storageKey": null
+    },
+    {
+      "alias": null,
+      "args": null,
+      "kind": "ScalarField",
+      "name": "name",
+      "storageKey": null
+    },
+    {
+      "alias": null,
+      "args": null,
+      "concreteType": "Image",
+      "kind": "LinkedField",
+      "name": "imageCover",
+      "plural": false,
+      "selections": [
+        {
+          "alias": null,
+          "args": null,
+          "kind": "ScalarField",
+          "name": "blurhash",
+          "storageKey": null
+        },
+        {
+          "alias": null,
+          "args": null,
+          "kind": "ScalarField",
+          "name": "url",
+          "storageKey": null
+        }
+      ],
+      "storageKey": null
+    },
+    {
+      "alias": null,
+      "args": null,
+      "concreteType": "Price",
+      "kind": "LinkedField",
+      "name": "price",
+      "plural": false,
+      "selections": [
+        {
+          "alias": null,
+          "args": null,
+          "kind": "ScalarField",
+          "name": "unitAmount",
+          "storageKey": null
+        },
+        {
+          "alias": null,
+          "args": null,
+          "kind": "ScalarField",
+          "name": "unitAmountCurrency",
+          "storageKey": null
+        }
+      ],
+      "storageKey": null
+    }
+  ],
+  "type": "Product",
+  "abstractKey": null
+};
+// prettier-ignore
+(node: any).hash = '9bc1cea9219bccd287fa9f0095e64d89';
+export default node;